requirements.txt
opencv-python~=4.6.0.66
numpy~=1.22.4
Pillow~=8.2.0
progress~=1.6
pandas~=1.2.4
numba~=0.53.1
matplotlib~=3.4.1